+++ packages/cubejs-backend-native/src/python/runtime.rs
@@ -1,7 +1,9 @@
 use crate::cross::CLRepr;
 use crate::python::neon_py::*;
+use crate::python::utils::PyAnyHelpers;
 use crate::tokio_runtime_node;
 use cubesql::CubeError;
+use futures::FutureExt;
 use log::{error, trace};
 use neon::prelude::*;
 use neon::types::Deferred;
@@ -10,6 +12,7 @@ use pyo3::prelude::*;
 use pyo3::types::{PyFunction, PyTuple};
 use std::fmt::Formatter;
 use std::future::Future;
+use std::panic;
 use std::pin::Pin;
 
 #[derive(Debug)]
@@ -46,6 +49,7 @@ enum PyScheduledFunResult {
 
 pub struct PyRuntime {
     sender: tokio::sync::mpsc::Sender<PyScheduledFun>,
+    js_channel: neon::event::Channel,
 }
 
 impl PyRuntime {
@@ -55,16 +59,49 @@ impl PyRuntime {
         args: Vec<CLRepr>,
         deferred: Deferred,
     ) {
-        let res = self.sender.blocking_send(PyScheduledFun {
+        // Try to reserve immediately for the fast path
+        let permit = match self.sender.try_reserve() {
+            Ok(permit) => permit,
+            Err(tokio::sync::mpsc::error::TrySendError::Full(_)) => {
+                log::warn!("Python channel is full, this may cause performance issues. Consider increasing the channel size for PyRuntime.");
+
+                // Channel is full, use async reserve with blocking for efficiency
+                match futures::executor::block_on(self.sender.reserve()) {
+                    Ok(permit) => permit,
+                    Err(_) => {
+                        // Channel was closed while waiting
+                        deferred.settle_with(
+                            &self.js_channel,
+                            move |mut cx| -> NeonResult<Handle<JsError>> {
+                                cx.throw_error(
+                                    "Unable to schedule python function call: channel is closed",
+                                )
+                            },
+                        );
+                        return;
+                    }
+                }
+            }
+            Err(tokio::sync::mpsc::error::TrySendError::Closed(_)) => {
+                // Channel is closed, settle deferred with error
+                deferred.settle_with(
+                    &self.js_channel,
+                    move |mut cx| -> NeonResult<Handle<JsError>> {
+                        cx.throw_error("Unable to schedule python function call: channel is closed")
+                    },
+                );
+                return;
+            }
+        };
+
+        let scheduled_fun = PyScheduledFun {
             fun,
             args,
             callback: PyScheduledCallback::NodeDeferred(deferred),
-        });
-        if let Err(err) = res {
-            // TODO: We need to return this error to deferred, but for now
-            // neon will handle this issue on Drop
-            error!("Unable to schedule python function call: {}", err)
-        }
+        };
+
+        // This should never fail since we have a permit
+        permit.send(scheduled_fun);
     }
 
     pub async fn call_async(
@@ -94,31 +131,42 @@ impl PyRuntime {
     ) -> Result<(), CubeError> {
         let (fun, args, callback) = task.split();
 
-        let task_result = Python::with_gil(move |py| -> PyResult<PyScheduledFunResult> {
-            let mut prep_tuple = Vec::with_capacity(args.len());
-            let mut py_kwargs = None;
+        let task_block = panic::AssertUnwindSafe(|| {
+            Python::with_gil(move |py| -> PyResult<PyScheduledFunResult> {
+                let mut prep_tuple = Vec::with_capacity(args.len());
+                let mut py_kwargs = None;
 
-            for arg in args {
-                if arg.is_kwarg() {
-                    py_kwargs = Some(arg.into_py_dict(py)?);
-                } else {
-                    prep_tuple.push(arg.into_py(py)?);
+                for arg in args {
+                    if arg.is_kwarg() {
+                        py_kwargs = Some(arg.into_py_dict(py)?);
+                    } else {
+                        prep_tuple.push(arg.into_py(py)?);
+                    }
                 }
-            }
 
-            let py_args = PyTuple::new(py, prep_tuple);
-            let call_res = fun.call(py, py_args, py_kwargs)?;
+                let py_args = PyTuple::new(py, prep_tuple);
+                let call_res = fun.call(py, py_args, py_kwargs)?;
 
-            let is_coroutine = unsafe { pyo3::ffi::PyCoro_CheckExact(call_res.as_ptr()) == 1 };
-            if is_coroutine {
-                let fut = pyo3_asyncio::tokio::into_future(call_res.as_ref(py))?;
-                Ok(PyScheduledFunResult::Poll(Box::pin(fut)))
-            } else {
-                Ok(PyScheduledFunResult::Ready(CLRepr::from_python_ref(
-                    call_res.as_ref(py),
-                )?))
-            }
+                if call_res.is_coroutine()? {
+                    let fut = pyo3_asyncio::tokio::into_future(call_res.as_ref(py))?;
+                    Ok(PyScheduledFunResult::Poll(Box::pin(fut)))
+                } else {
+                    Ok(PyScheduledFunResult::Ready(CLRepr::from_python_ref(
+                        call_res.as_ref(py),
+                    )?))
+                }
+            })
         });
+
+        let task_result = match panic::catch_unwind(task_block) {
+            Ok(Ok(r)) => Ok(r),
+            Ok(Err(err)) => Err(CubeError::user(format_python_error(err))),
+            Err(panic_payload) => Err(CubeError::panic_with_message(
+                panic_payload,
+                "Unexpected panic while calling python function",
+            )),
+        };
+
         let task_result = match task_result {
             Ok(r) => r,
             Err(err) => {
@@ -127,13 +175,12 @@ impl PyRuntime {
                         deferred.settle_with(
                             js_channel,
                             move |mut cx| -> NeonResult<Handle<JsError>> {
-                                cx.throw_from_python_error(err)
+                                cx.throw_error(err.to_string())
                             },
                         );
                     }
                     PyScheduledCallback::Channel(chan) => {
-                        let send_res =
-                            chan.send(Err(CubeError::internal(format_python_error(err))));
+                        let send_res = chan.send(Err(err));
                         if send_res.is_err() {
                             return Err(CubeError::internal(
                                 "Unable to send result back to consumer".to_string(),
@@ -151,31 +198,37 @@ impl PyRuntime {
                 let js_channel_to_move = js_channel.clone();
 
                 tokio::spawn(async move {
-                    let fut_res = fut.await;
+                    let safe_py_fut_poll = panic::AssertUnwindSafe(async {
+                        let fut_res = fut.await;
 
-                    let res = Python::with_gil(move |py| -> Result<CLRepr, PyErr> {
-                        let res = match fut_res {
-                            Ok(r) => CLRepr::from_python_ref(r.as_ref(py)),
-                            Err(err) => Err(err),
-                        };
+                        Python::with_gil(move |py| -> Result<CLRepr, PyErr> {
+                            let res = match fut_res {
+                                Ok(r) => CLRepr::from_python_ref(r.as_ref(py)),
+                                Err(err) => Err(err),
+                            };
 
-                        res
+                            res
+                        })
                     });
+
+                    let fut_res = match safe_py_fut_poll.catch_unwind().await {
+                        Ok(Ok(r)) => Ok(r),
+                        Ok(Err(err)) => Err(CubeError::internal(format_python_error(err))),
+                        Err(panic_payload) => Err(CubeError::panic_with_message(
+                            panic_payload,
+                            "Unexpected panic while polling python future",
+                        )),
+                    };
 
                     match callback {
                         PyScheduledCallback::NodeDeferred(deferred) => {
-                            deferred.settle_with(&js_channel_to_move, |mut cx| match res {
-                                Err(err) => cx.throw_error(format!("Python error: {}", err)),
+                            deferred.settle_with(&js_channel_to_move, |mut cx| match fut_res {
                                 Ok(r) => r.into_js(&mut cx),
+                                Err(err) => cx.throw_error(format!("{}", err)),
                             });
                         }
                         PyScheduledCallback::Channel(chan) => {
-                            let _ = match res {
-                                Ok(r) => chan.send(Ok(r)),
-                                Err(err) => {
-                                    chan.send(Err(CubeError::internal(format_python_error(err))))
-                                }
-                            };
+                            let _ = chan.send(fut_res);
                         }
                     }
                 });
@@ -202,7 +255,9 @@ impl PyRuntime {
 
         trace!("New Python runtime");
 
-        std::thread::spawn(|| {
+        let js_channel_clone = js_channel.clone();
+
+        std::thread::spawn(move || {
             trace!("Initializing executor in a separate thread");
 
             std::thread::spawn(|| {
@@ -216,7 +271,7 @@ impl PyRuntime {
                         if let Some(task) = receiver.recv().await {
                             trace!("New task");
 
-                            if let Err(err) = Self::process_task(task, &js_channel) {
+                            if let Err(err) = Self::process_task(task, &js_channel_clone) {
                                 error!("Error while processing python task: {:?}", err)
                             };
                         }
@@ -229,7 +284,7 @@ impl PyRuntime {
             }
         });
 
-        Self { sender }
+        Self { sender, js_channel }
     }
 }