+++ rust/cubesql/cubesql/benches/large_model.rs
@@ -73,6 +73,7 @@ pub fn get_large_model_test_meta(dims: usize) -> Vec<V1CubeMeta> {
                 agg_type: Some("count".to_string()),
                 meta: None,
                 alias_member: None,
+                format: None,
             },
             V1CubeMetaMeasure {
                 name: format!("{}.sum", cube_name),
@@ -83,6 +84,7 @@ pub fn get_large_model_test_meta(dims: usize) -> Vec<V1CubeMeta> {
                 agg_type: Some("sum".to_string()),
                 meta: None,
                 alias_member: None,
+                format: None,
             },
         ],
         dimensions: (1..=dims)